Refactor generic Pptr utils into own file
