+++ tests/e2e/utils.js
@@ -11,7 +11,11 @@ import {
 import { WP_ADMIN_DASHBOARD } from '@woocommerce/e2e-utils';
 
 /**
- * @typedef {import('@types/puppeteer').Page} Page
+ * Internal dependencies
+ */
+import { elementExists, getTextContent } from './page-utils';
+
+/**
  * @typedef {import('@types/puppeteer').ElementHandle} ElementHandle
  * @typedef {import('@wordpress/blocks').Block} WPBlock
  */
@@ -149,40 +153,11 @@ export async function goToSiteEditor( query, editorContext = 'core' ) {
  * Waits for the Gutenberg canvas to be available
  *
  * @param {number} [timeout=DEFAULT_TIMEOUT] The amount of ms to wait for the element
+ *
+ * @return {Promise<?ElementHandle>} The canvas element handle
  */
-export async function waitForCanvas( timeout = DEFAULT_TIMEOUT ) {
-	await page.waitForSelector( SELECTORS.canvas, { timeout } );
-}
-
-/**
- * Gets the text value of an element
- *
- * If the element is an `input` it will get the `value`, otherwise,
- * it will get the `textContent`.
- *
- * @param {string} selector The selector for the desired element
- * @param {Page | ElementHandle} [root=page] The root from which to search for the selector
- *
- * @return {Promise<string[]>} An array of text contained in those selected elements
- */
-export async function getTextContent( selector, root = page ) {
-	return root.$$eval( selector, ( $elements ) => {
-		return $elements.map(
-			( $element ) => $element.value || $element.textContent
-		);
-	} );
-}
-
-/**
- * Checks whether an element exists under a certain context
- *
- * @param {string} selector The selector for the desired element
- * @param {Page | ElementHandle} [root=page] The root from which to search for the selector
- *
- * @return {Promise<boolean>} Whether the element exists or not
- */
-export async function elementExists( selector, root = page ) {
-	return !! ( await root.$( selector ) );
+export function waitForCanvas( timeout = DEFAULT_TIMEOUT ) {
+	return page.waitForSelector( SELECTORS.canvas, { timeout } );
 }
 
 /**
